requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
git-filter-repo==2.47.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
numpy==2.3.0
oauth2client==4.1.3
oauthlib==3.3.1
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyMySQL==1.1.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dotenv==1.1.1
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tesseract==0.1.3
uritemplate==4.2.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
